REFACTOR: Rich text in own file
